requirements.txt
python-telegram-bot==13.15
vk_api==11.10.0
redis==7.0.1
python-dotenv==1.2.1
